Changes to HTML and CSS
Added in actual content for the webpage including images for the skills and real descriptions. CSS was further refined to provide a much more palatable website. <hr> was given style, the skill rows were given floats and clears, the footer was given style and formatting. A button was added to create a method to contact me.