Modified index.html and styles.css
Index.html saw a change in the class of the button from "contact-me-btn" to just "btn." Styles.css saw the line-height property move to the "body" tag. The paddings for the containers, paragraphs, and anchor tags was changed slightly. "a:hover" was added in, as well as "font-style" for the anchor tags.